tests: migrate show_graph tests from System->Composition
